move HepMCFilters back into GeneratorInterface/Core and remove new package
