+++ GeneratorInterface/Core/BuildFile.xml
@@ -3,7 +3,6 @@
 <use   name="FWCore/Utilities"/>
 <use   name="SimDataFormats/GeneratorProducts"/>
 <use   name="GeneratorInterface/LHEInterface"/>
-<use   name="GeneratorInterface/HepMCFilters"/>
 <use   name="boost"/>
 <use   name="clhep"/>
 <use   name="lhapdf"/>
