--- GeneratorInterface/HepMCFilters/BuildFile.xml
@@ -1,6 +1,0 @@
-<use   name="FWCore/Framework"/>
-<use   name="SimDataFormats/GeneratorProducts"/>
-<use   name="clhep"/>
-<export>
-  <lib   name="1"/>
-</export>
